Remove useBoilerHouseDetails
